New users are granted default permissions
There are 3 places a user can be created:

1. the /users form
2. the /batch_invitations form
3. the users:create rake task

None of which use exactly the same mechanism so we have to introduce the
code in each place.  It may be worth refactoring so that all three places
use the same user creation service object (perhaps we can extend the
one we extracted from the rake task).  One obvious alternative would be
to use an after_create callback, but I'm wary of that as callbacks can
end up getting expensive.  Particularly in tests.