Add processed and total user counts to bulk grant permission set
These counters allow us to track progress during the perform method in
a similar fashion to how the batch invitations does it with it's user
association.

In rare circumstances we might end up with processed users > total users
if someone is creats a user while our job is processing, but it's not a
huge problem.  If it is we might want to fix the user ids we're going to
process into the permission set at creation time.  This would pave the
way towards us being able to choose a subset of users to bulk grant to,
which might be interesting, but we don't need it now.