Add default flag to supported permissions
This flag indicates if this is a permission that should be granted by
default to all new users.  For example, we want all users to be able
to use the support application, so rather than asking the person creating
the user to remember to give them support permissions we should be able
to flag it as default once and have the system grant it for us.

This commit adds the flag and allows admins to edit permissions to say
which are default or not.  Future commits will change the system to act
on this flag when creating users.