Remove special case for support app on dashboard
The dashboard would always provide a link to the support app regardless
of the permissions the user has on the support app.  This was because the
support app didn't explicitly require "signin" permissions and we wanted
everyone to have a link to it.  Now that signon handles the 'signin'
permission itself this link wouldn't always work and would confuse users
who didn't really have the permission so we remove it.  To keep the
existing experience though we'll be giving everyone "signin" permission
via the bulk grant functionality and making it a default permission for
new users.